removed duplicate file structure. Steve had me include a workaround for the url since ClassLoader wasn't working
