--- ManualHttpServer/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Wed Jul 18 13:46:55 PDT 2018
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-3.5-rc-2-bin.zip
