--- ManualHttpServer/src/main/resources/static/quote.html
@@ -1,9 +1,0 @@
-<html>
-  <head>
-    <title>Quotes</title>
-  </head>
-  <body>
-    <h1>A Wise Man Once Said</h1>
-    <p>{{RANDOM_JSON_QUOTE}}</p>
-  </body>
-</html>
