--- ManualHttpServer/src/main/resources/static/seattle.html
@@ -1,10 +1,0 @@
-<html>
-  <head>
-    <title>Home</title>
-  </head>
-  <body>
-    <h1>Seattle! Ahhhh, Home</h1>
-    <p>Seattle, a wonderful place to be. Especially in the summer</p>
-    <p>Nothing better than the summer in the northwest.</p>
-  </body>
-</html>
