--- ManualHttpServer/static/index.html
@@ -1,15 +1,0 @@
-<html>
-  <head>
-    <title>Hello</title>
-  </head>
-  <body>
-    <h1>So cool</h1>
-    <p>The current time is {{TIMESTAMP}}</p>
-
-    <ul>
-      <li>
-        check out <a href="/seattle.html">seattle</a>
-      </li>
-    </ul>
-  </body>
-</html>
