--- ManualHttpServer/static/madlibs.html
@@ -1,15 +1,0 @@
-<html>
-  <head>
-    <title>Hello</title>
-  </head>
-  <body>
-    <h1>So cool</h1>
-    <pre>
-        {{LIB}}! he said
-        {{LIB}} as he jumped into his convertible
-        exclamation            adverb
-        {{LIB}} and drove off with his {{LIB}} friend."
-        noun                          adjective
-    </pre>
-  </body>
-</html>
