+++ src/main/java/TemplateFileReader.java
@@ -1,3 +1,4 @@
+
 import java.util.Map;
 
 public class TemplateFileReader extends HTTPStaticFileReader {
