+++ src/main/resources/quote.html
